move Xliff functions
