+++ extension/src/test/XlfHighlighter.test.ts
@@ -2,7 +2,6 @@ import * as vscode from "vscode";
 import * as path from "path";
 import * as XlfHighlighter from "../XlfHighlighter";
 import * as assert from "assert";
-import * as LanguageFunctions from "../LanguageFunctions";
 import { LanguageFunctionsSettings } from "../Settings/LanguageFunctionsSettings";
 import {
   invalidXmlSearchExpression,
@@ -10,6 +9,7 @@ import {
 } from "../constants";
 import * as SettingsLoader from "../Settings/SettingsLoader";
 import { TranslationMode } from "../Enums";
+import * as XliffFunctions from "../XliffFunctions";
 
 const testResourcesPath = "../../src/test/resources/highlights/";
 const translationTokenXlfUri: vscode.Uri = vscode.Uri.file(
@@ -62,7 +62,7 @@ suite("Xlf Highlighter", function () {
 
     await assert.rejects(
       async () => {
-        await LanguageFunctions._refreshXlfFilesFromGXlf({
+        await XliffFunctions._refreshXlfFilesFromGXlf({
           gXlfFilePath: gXlfUri,
           langFiles: langFilesUri,
           languageFunctionsSettings,
